backend/requirements.txt
itsdangerous>=2.0.0
starlette>=0.27.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
authlib==1.2.1
requests==2.31.0
kubernetes==28.1.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
asyncio-mqtt==0.16.1
structlog==23.2.0
cryptography==41.0.7
aiohttp==3.9.1

# Security dependencies for enhanced authentication
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
redis>=4.5.0  # For distributed rate limiting and session storage

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.27.0  # For TestClient and OAuth client

# Cloud Provider dependencies
# GCP storage management
google-cloud-storage==2.10.0

# AWS storage management (optional - only needed for AWS deployments)
boto3>=1.34.0  # AWS SDK for Python